Added different logging levels, and error selection for new machine
Added different logging levels dependant on the logggingLevel variable,
Added the ability to enter a comma seperated value in json during the creation of a new machine to specify which errors will be being reported. all errors are setup to be nullable meaning adding the wrong error will still allow short time statistics packets to work however there will be an unused column in the table.