Lots of commenting
Heading out for lunch but will continue when i get back. Primarily just ocmmenting the hell out of the code to make it nice and easy to read for even those who dont know c# ( most here do not i want it easy to maintain once im gone)
Also added more to the App.config as the less they have to open this the better.